+++ src/workspace/toolbar.tsx
@@ -1,4 +1,5 @@
 import * as React from 'react';
+import classnames from 'classnames';
 
 import { WorkspaceLanguage } from './workspace';
 
@@ -24,9 +25,88 @@ export interface ToolbarProps {
 const CLASS_NAME = 'ontodia-toolbar';
 
 export class DefaultToolbar extends React.Component<ToolbarProps, {}> {
-    private onChangeLanguage = (event: React.SyntheticEvent<HTMLSelectElement>) => {
-        const value = event.currentTarget.value;
-        this.props.onChangeLanguage?.(value);
+    render() {
+        return (
+            <div className={CLASS_NAME}>
+                <div className='ontodia-btn-group ontodia-btn-group-sm'>
+                    {this.renderSaveDiagramButton()}
+                    {this.renderPersistAuthoredChangesButton()}
+                    {this.props.onClearAll ? (
+                        <button type='button'
+                            className={classnames(
+                                `${CLASS_NAME}__clear-all-button`,
+                                'ontodia-btn ontodia-btn-default'
+                            )}
+                            title='Clear All' onClick={this.props.onClearAll}>
+                            Clear All
+                        </button>
+                    ) : null}
+                    <button type='button'
+                        className={classnames(
+                            `${CLASS_NAME}__layout-button`,
+                            'ontodia-btn ontodia-btn-default'
+                        )}
+                        title='Force layout' onClick={this.props.onForceLayout}>
+                         Layout
+                    </button>
+                    <button type='button'
+                        className={classnames(
+                            `${CLASS_NAME}__export-button`,
+                            'ontodia-btn ontodia-btn-default'
+                        )}
+                        title='Export diagram as PNG' onClick={this.onExportPNG}>
+                        PNG
+                    </button>
+                    <button type='button'
+                        className={classnames(
+                            `${CLASS_NAME}__export-button`,
+                            'ontodia-btn ontodia-btn-default'
+                        )}
+                        title='Export diagram as SVG' onClick={this.onExportSVG}>
+                        SVG
+                    </button>
+                    <button type='button'
+                        className={classnames(
+                            `${CLASS_NAME}__print-button`,
+                            'ontodia-btn ontodia-btn-default'
+                        )}
+                        title='Print diagram' onClick={this.props.onPrint}>
+                        Print
+                    </button>
+                    {this.renderLanguages()}
+                </div>
+            </div>
+        );
+    }
+
+    private renderSaveDiagramButton() {
+        if (!this.props.onSaveDiagram) { return null; }
+        return (
+            <button type='button'
+                className={classnames(
+                    `${CLASS_NAME}__save-button`,
+                    'ontodia-btn ontodia-btn-primary'
+                )}
+                disabled={this.props.canSaveDiagram === false}
+                onClick={this.props.onSaveDiagram}>
+                Save diagram
+            </button>
+        );
+    }
+
+    private renderPersistAuthoredChangesButton() {
+        if (!this.props.onPersistChanges) { return null; }
+        return (
+            <button type='button'
+                className={classnames(
+                    `${CLASS_NAME}__save-button`,
+                    'ontodia-btn ontodia-btn-success'
+                )}
+                disabled={this.props.canPersistChanges === false}
+                onClick={this.props.onPersistChanges}>
+                Save data
+            </button>
+        );
     }
 
     private onExportSVG = () => {
@@ -37,35 +117,13 @@ export class DefaultToolbar extends React.Component<ToolbarProps, {}> {
         this.props.onExportPNG?.();
     }
 
-    private renderSaveDiagramButton() {
-        if (!this.props.onSaveDiagram) { return null; }
-        return (
-            <button type='button' className='saveDiagramButton ontodia-btn ontodia-btn-primary'
-                disabled={this.props.canSaveDiagram === false}
-                onClick={this.props.onSaveDiagram}>
-                <span className='fa fa-floppy-o' aria-hidden='true' /> Save diagram
-            </button>
-        );
-    }
-
-    private renderPersistAuthoredChangesButton() {
-        if (!this.props.onPersistChanges) { return null; }
-        return (
-            <button type='button' className='saveDiagramButton ontodia-btn ontodia-btn-success'
-                disabled={this.props.canPersistChanges === false}
-                onClick={this.props.onPersistChanges}>
-                <span className='fa fa-floppy-o' aria-hidden='true' /> Save data
-            </button>
-        );
-    }
-
     private renderLanguages() {
         const {selectedLanguage, languages} = this.props;
         if (!(languages && languages.length > 0)) { return null; }
 
         return (
-            <span className={`ontodia-btn-group ${CLASS_NAME}__language-selector`}>
-                <label className='ontodia-label'><span>Data Language - </span></label>
+            <span className={`${CLASS_NAME}__language-selector`}>
+                <label className='ontodia-label'><span>Data Language:</span></label>
                 <select value={selectedLanguage} onChange={this.onChangeLanguage}>
                     {languages.map(({code, label}) => <option key={code} value={code}>{label}</option>)}
                 </select>
@@ -73,49 +131,8 @@ export class DefaultToolbar extends React.Component<ToolbarProps, {}> {
         );
     }
 
-    render() {
-        return (
-            <div className={CLASS_NAME}>
-                <div className='ontodia-btn-group ontodia-btn-group-sm'>
-                    {this.renderSaveDiagramButton()}
-                    {this.renderPersistAuthoredChangesButton()}
-                    {this.props.onClearAll ? (
-                        <button type='button' className='ontodia-btn ontodia-btn-default'
-                            title='Clear All' onClick={this.props.onClearAll}>
-                            <span className='fa fa-trash' aria-hidden='true'/>&nbsp;Clear All
-                        </button>
-                    ) : null}
-                    <button type='button' className='ontodia-btn ontodia-btn-default'
-                            title='Force layout' onClick={this.props.onForceLayout}>
-                        <span className='fa fa-sitemap' aria-hidden='true'/> Layout
-                    </button>
-                    <button type='button' className='ontodia-btn ontodia-btn-default'
-                            title='Zoom In' onClick={this.props.onZoomIn}>
-                        <span className='fa fa-search-plus' aria-hidden='true'/>
-                    </button>
-                    <button type='button' className='ontodia-btn ontodia-btn-default'
-                            title='Zoom Out' onClick={this.props.onZoomOut}>
-                        <span className='fa fa-search-minus' aria-hidden='true'/>
-                    </button>
-                    <button type='button' className='ontodia-btn ontodia-btn-default'
-                            title='Fit to Screen' onClick={this.props.onZoomToFit}>
-                        <span className='fa fa-arrows-alt' aria-hidden='true'/>
-                    </button>
-                    <button type='button' className='ontodia-btn ontodia-btn-default'
-                            title='Export diagram as PNG' onClick={this.onExportPNG}>
-                        <span className='fa fa-picture-o' aria-hidden='true'/> PNG
-                    </button>
-                    <button type='button' className='ontodia-btn ontodia-btn-default'
-                            title='Export diagram as SVG' onClick={this.onExportSVG}>
-                        <span className='fa fa-picture-o' aria-hidden='true'/> SVG
-                    </button>
-                    <button type='button' className='ontodia-btn ontodia-btn-default'
-                            title='Print diagram' onClick={this.props.onPrint}>
-                        <span className='fa fa-print' aria-hidden='true'/>
-                    </button>
-                    {this.renderLanguages()}
-                </div>
-            </div>
-        );
+    private onChangeLanguage = (event: React.SyntheticEvent<HTMLSelectElement>) => {
+        const value = event.currentTarget.value;
+        this.props.onChangeLanguage?.(value);
     }
 }